+++ CHANGELOG.rst
@@ -2,6 +2,7 @@ Changelog
 =========
 
 - Fix semiparametric synthetic likelihood with glasso/warton and add tests
+- Improve batch simulations in toad example
 - Remove synthetic likelihood node and update BSL data collection
 - Fix M/G/1 example
 - Fix scratch assay example